+++ src/shadergen/src/DxcShaderCompiler.h
@@ -21,7 +21,7 @@ namespace sg
                           const std::string& filePath,
                           const char* entryPoint,
                           uint32_t* spvSize,
-                          uint32_t** spv) override;
+                          uint8_t** spv) override;
 
   private:
     CComPtr<IDxcCompiler3> m_compiler;
